tox: add py to dependencies
pytest dropped `py` as dependency and pytest-xprocess requires it as
implicit dependency. As pytest-xprocess relied on pytest to pull it in
it now errors when trying to run the murdock tests.

This commit adds `py` to the list of dependencies for tox. It can be
reverted as soon as pytest-xprocess adds `py` to the dependency list.
See also: https://github.com/pytest-dev/pytest-xprocess/issues/110