murdock: Add prometheus counters for jobs
This commit adds counters for the number of jobs added to the queue and
the result status of jobs.

Any job (re-)entering the queue should be counted, intentionally also
counting jobs that were requeued because of a change in the pull
request.

The other counter keeps track of the result of a job. This includes the
passed and errored label if the job completed the build, but also
stopped and canceled if the job was stopped before completing or if the
job was not started yet.

The number of builds currently in the queue should match the difference
between the queue counter and the job status counter