Rewrite logging to structured log format
This commit rewrites the logging in the application to structured log
format. A pretty console printing and direct systemd-journal are available.
The logging can be switched via the MURDOCK_LOG_OUTPUT environment
variable. It defaults to `console` for development. Setting it to
`journal` sets it to output to the systemd journal.