github: don't log full comment response from github
This removes the content of the response from github after posting a
reply to a topic. It still logs the content after an error. In case of
success the full json of the post is in the reply which is quite a lot
of data to log while it doesn't add anything